+++ Server/routes/index.js
@@ -1,179 +1,12 @@
 var express = require('express');
 var router = express.Router();
-var read = require('node-readability');
-var jaccard = require('multiset-jaccard');
-var mecab = require('mecab-ya');
-var textrank = require('textrank-node');
-var ranker = new textrank();
-var async = require('async');
-
-
-/* 리팩토링 1. 텍스트를 문장/ 문단으로 분리 2. 그래프를 얻는다 3. 텍스트랭크 돌린다.
- 파싱한 텍스트 다듬기
-1. () 와 안의 내용물은 제거한다
-2. 온점 없는 문장은 제거한다. => 개행문자로 분리 , 온점 없으면 삭제. 필터링
-3.
-*/
-var emailRegex = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
-
-function refineText(text, next) {
-  async.waterfall([
-    function(callback){ // 필요없는 부분 제거. (), 이메일 제거 등등..
-      text = text.replace(/\([^)]+\)/g, '');
-      text = text.replace(/\[[^\]]+\]/g, '');
-      text = text.replace(/\S+@\S+\.\S+/, ''); // 이메일 제거
-      text = text.replace(/\S+@\S+\.\S+\.\S+/, '');
-      callback(null, text);
-    },
-    function(text, callback) { // 문단으로 분리
-      splitToParagraphs(text, callback);
-    },
-    function(paragraphs, callback){ // 온점이 없는 문단은 지운다.
-      paragraphs = paragraphs.filter(function(para) {
-        if(para.includes('.')) return true;
-        else return false;
-      });
-      callback(null, paragraphs);
-    },
-    function(paragraphs, callback) { // 각 문단을 문장으로 쪼개기서 하나의 배열에 넣기.
-      /// 여기까진 되고..
-      splitToSentences (paragraphs, callback);
-    }
-  ],function(err, result){
-    if(err) throw err;
-    next(result);
-  });
-}
-
-function splitToParagraphs (text, callback) {
-  var paragraphs = text.split('\n');
-  callback(null, paragraphs);
-}
-
-function splitParagraphsSentences (paragraphs, callback) {
-  var result = [];
-  async.eachSeries(paragraphs, function(para, callback){
-    var sentences = para.match(/[^\.!\?\/]+[\.!\?]+/g);
-    if(sentences.length !== 0){
-      result.push(sentences);
-    }
-    callback();
-  }, function(err){
-    callback(null, result);
-  });
-}
-
-function splitToSentences (paragraphs, callback) {
-  var result = [];
-  async.eachSeries(paragraphs, function(para, callback){
-    var sentences = para.match(/[^\.!\?\/]+[\.!\?]+/g);
-    if(sentences.length !== 0){
-      for(var i = 0; i < sentences.length; i++){
-        result.push(sentences[i]);
-      }
-    }
-    callback();
-  }, function(err){
-    console.log(result);
-    callback(null, result);
-  });
-}
-
-
-// 이건 하나의 문단 (문장들의 배열) 안에서 그래프를 구하는 것임
-function getGraph (lines, callback) {
-  var graph = [];
-  // 1st para in async.each() is the array of items
-  async.eachSeries(lines, function(leftLine, finishLeft){
-      // Call an asynchronous function, often a save() to DB
-      var sentenceSimilarity = [];
-      mecab.nouns(leftLine, function(err, leftNouns) {
-        async.eachSeries(lines, function(rightLine, finishRight) {
-          mecab.nouns(rightLine, function(err, rightNouns) {
-            if(leftNouns === undefined)
-            { leftNouns = [""]; }
-            if(rightNouns === undefined)
-            { rightNouns = [""]; }
-
-            jaccard.index(leftNouns, rightNouns, function(err, index) {
-              sentenceSimilarity.push(index);
-              finishRight();
-            });
-          });
-        }, function(err) {
-          graph.push(sentenceSimilarity);
-          finishLeft();
-        });
-      });
-    },
-    // 3rd param is the function to call when everything's done
-    function(err){
-      callback(err, graph);
-    }
-  );
-}
-
-function getSelectedLines(lines, selectedIndex, callback){
-  // 문장의 순서대로 정렬하는 게 문맥상 자연스러운 듯.
-  async.sortBy(selectedIndex, function(x, callback) {
-    callback(null, x);
-  }, function(err, sortedIndex) {
-    var result = [];
-    for(var i = 0; i < 3; i++){
-      for(var j = 0; j < lines.length; j++){
-        if(sortedIndex[i] === j){
-          result[i] = lines[j];
-        }
-      }
-    }
-    callback(result);
-  });
-}
-
-function getSelectedIndex(textRank, max) {
-  if(textRank.length === 0){
-    return [0];
-  }
-  var selectedIndex = [];
-
-  for(var i = 0; i < max&& i < textRank.length; i++){
-    selectedIndex[i] = textRank.length - (i+1); // 매번 같은 인덱스를 기본으로 두면 문제가 생김.
-    for(var j = 0; j < textRank.length; j++){
-      if(textRank[selectedIndex[i]] < textRank[j]){
-        if(!selectedIndex.includes(j)){
-          selectedIndex[i] = j;
-        }
-      }
-    }
-  }
-  return selectedIndex;
-}
-
+var readability = require('../readability');
+var summarize = require('../summarize');
 
 router.get('/hello', function (req, res, next) {
-  read(req.headers.pageurl, function(err, article, meta) {
-
-    var text = "";
-    if(req.headers.pageurl.includes("news.naver")){
-      var dom = article.document;
-      //<br> 사이에 있는 것들만 각각 문단으로 가져오게 하기
-      text = dom.getElementById("articleBodyContents").textContent;
-    }
-    else{
-      text = article.textBody;
-    }
-    console.log(text);
-    //var lines = splitToSentences(text);
-    refineText(text, function(lines) {
-        getGraph(lines, function(err, graph) {
-          // 그래프를 얻었으니 텍스트랭크를 돌린다.
-          var Rank = ranker.getTextRank(graph).probabilityNodes;
-          // 가장 영향력이 큰 노드 세 개를 취한다.
-          var selectedIndex = getSelectedIndex(Rank, 3);
-          getSelectedLines(lines, selectedIndex, function(result){
-            return res.json({shorten: result});
-          });
-        });
+  readability.refine(req.headers.pageurl, function(lines) {
+    summarize.summarize(lines, function(summarized){
+      return res.json({ shorten: summarized});
     });
   });
 });
